[Maps] provide function to create basemap layer descriptor (#85617)
* [Maps] provide function to create basemap layer descriptor

* update typing

* update APM map to use createBasemapLayerDescriptor

* fix useLayerList jest test

* tslint fixes
