+++ x-pack/plugins/apm/public/components/app/RumDashboard/VisitorBreakdownMap/EmbeddedMap.tsx
@@ -22,6 +22,7 @@ import {
 } from '../../../../../../../../src/plugins/embeddable/public';
 import { useLayerList } from './useLayerList';
 import { useUrlParams } from '../../../../context/url_params_context/use_url_params';
+import { useApmPluginContext } from '../../../../context/apm_plugin/use_apm_plugin_context';
 import { RenderTooltipContentParams } from '../../../../../../maps/public';
 import { MapToolTip } from './MapToolTip';
 import { useMapFilters } from './useMapFilters';
@@ -50,6 +51,7 @@ interface KibanaDeps {
 }
 export function EmbeddedMapComponent() {
   const { urlParams } = useUrlParams();
+  const apmPluginContext = useApmPluginContext();
 
   const { start, end, serviceName } = urlParams;
 
@@ -151,6 +153,12 @@ export function EmbeddedMapComponent() {
 
       if (embeddableObject && !isErrorEmbeddable(embeddableObject)) {
         embeddableObject.setRenderTooltipContent(renderTooltipContent);
+        const basemapLayerDescriptor = apmPluginContext.plugins.maps
+          ? await apmPluginContext.plugins.maps.createLayerDescriptors.createBasemapLayerDescriptor()
+          : null;
+        if (basemapLayerDescriptor) {
+          layerList.unshift(basemapLayerDescriptor);
+        }
         await embeddableObject.setLayerList(layerList);
       }
