+++ x-pack/plugins/maps/public/api/index.ts
@@ -5,5 +5,5 @@
  */
 
 export { MapsStartApi } from './start_api';
-export { createSecurityLayerDescriptors } from './create_security_layer_descriptors';
+export { createLayerDescriptors } from './create_layer_descriptors';
 export { registerLayerWizard, registerSource } from './register';
