+++ x-pack/plugins/maps/public/lazy_load_bundle/lazy/index.ts
@@ -15,3 +15,4 @@ export { registerLayerWizard } from '../../classes/layers/layer_wizard_registry'
 export { registerSource } from '../../classes/sources/source_registry';
 export { createTileMapLayerDescriptor } from '../../classes/layers/create_tile_map_layer_descriptor';
 export { createRegionMapLayerDescriptor } from '../../classes/layers/create_region_map_layer_descriptor';
+export { createBasemapLayerDescriptor } from '../../classes/layers/create_basemap_layer_descriptor';
