+++ x-pack/plugins/maps/public/routes/map_page/saved_map/get_initial_layers_from_url_param.ts
@@ -4,7 +4,6 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import _ from 'lodash';
 import rison from 'rison-node';
 import { i18n } from '@kbn/i18n';
 import '../../../classes/sources/wms_source';
@@ -14,41 +13,8 @@ import '../../../classes/sources/es_pew_pew_source';
 import '../../../classes/sources/kibana_regionmap_source';
 import '../../../classes/sources/es_geo_grid_source';
 import '../../../classes/sources/xyz_tms_source';
-import { LayerDescriptor } from '../../../../common/descriptor_types';
-// @ts-expect-error
-import { KibanaTilemapSource } from '../../../classes/sources/kibana_tilemap_source';
-import { TileLayer } from '../../../classes/layers/tile_layer/tile_layer';
-// @ts-expect-error
-import { EMSTMSSource } from '../../../classes/sources/ems_tms_source';
-// @ts-expect-error
-import { VectorTileLayer } from '../../../classes/layers/vector_tile_layer/vector_tile_layer';
-import { getEMSSettings, getToasts } from '../../../kibana_services';
+import { getToasts } from '../../../kibana_services';
 import { INITIAL_LAYERS_KEY } from '../../../../common/constants';
-import { getKibanaTileMap } from '../../../meta';
-
-export function getInitialLayers(layerListJSON?: string, initialLayers: LayerDescriptor[] = []) {
-  if (layerListJSON) {
-    return JSON.parse(layerListJSON);
-  }
-
-  const tilemapSourceFromKibana = getKibanaTileMap();
-  if (_.get(tilemapSourceFromKibana, 'url')) {
-    const layerDescriptor = TileLayer.createDescriptor({
-      sourceDescriptor: KibanaTilemapSource.createDescriptor(),
-    });
-    return [layerDescriptor, ...initialLayers];
-  }
-
-  const isEmsEnabled = getEMSSettings()!.isEMSEnabled();
-  if (isEmsEnabled) {
-    const layerDescriptor = VectorTileLayer.createDescriptor({
-      sourceDescriptor: EMSTMSSource.createDescriptor({ isAutoSelect: true }),
-    });
-    return [layerDescriptor, ...initialLayers];
-  }
-
-  return initialLayers;
-}
 
 export function getInitialLayersFromUrlParam() {
   const locationSplit = window.location.href.split('?');
